Refactor RosCamera into RosInterface
